requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
jinja2
python-multipart
passlib[bcrypt]
pydantic
aiosmtplib
email-validator
python-jose[cryptography]
python-dotenv